refactor: apply-page 날짜 유틸 분리 및 스텝 이동 시 스크롤 최상단 고정
- formatKoreanDate → shared/utils/date-formatter.ts (일반 날짜 포맷 유틸)
- parseBirthDate, parseGraduationDate, parseLastSemester → widgets/apply/model/draft-parser.ts
  (apply-dropdown 의존성으로 shared 대신 widget 레이어에 배치)
- apply-page.tsx에서 인라인 정의 제거 및 임포트로 교체
- currentStep 변경 시 window.scrollTo로 화면 최상단 고정
